requirements.txt
django
djangorestframework
requests
redis
prettyconf
